Replace fragment with Compose dialog
